+++ test/distributed/pipelining/test_schedule.py
@@ -19,17 +19,19 @@ from torch.distributed.pipelining import (
     ScheduleZBVZeroBubble,
 )
 from torch.distributed.pipelining._utils import generate_stage_to_rank_mapping
-from torch.distributed.pipelining.schedules import (
-    _Action,
-    _add_reduce_grad,
+from torch.distributed.pipelining.runtime import (
     _add_send_recv,
     _add_unshard_reshard,
-    _format_pipeline_order,
     _merge_bw,
-    _PipelineSchedule,
     _PipelineScheduleRuntime,
     _simulate_comms_compute,
     _validate_schedule,
+)
+from torch.distributed.pipelining.schedules import (
+    _Action,
+    _add_reduce_grad,
+    _format_pipeline_order,
+    _PipelineSchedule,
     B,
     F,
     get_schedule_class,
